Fix a mismatching region id type

+++ include/Memory.hpp
@@ -13,7 +13,7 @@ namespace harava
 	{
 		memory_region(const std::string& range_str);
 		size_t start, end;
-		u64 id;
+		u16 id;
 
 		// if no results are from this region, ignore it
 		bool ignore = false;
